+++ Okta/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade sekoia-automation-sdk to 1.22.2
+- Upgrade AssetConnector to Async
 
 ## 2026-01-29 - 2.8.12